requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1